fix: change .eslintrc
