+++ controllers/movies.js
@@ -1,4 +1,5 @@
-const { NotFoundError, ForbiddenError } = require('../middlwares/errors');
+const { NotFoundError } = require('../errors/NotFoundError');
+const { ForbiddenError } = require('../errors/ForbiddenError');
 const Movie = require('../models/movie');
 
 // возвращает сохраненные пользователем фильмы
