+++ errors/BadRequestError.js
@@ -1,0 +1,8 @@
+class BadRequestError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 400;
+  }
+}
+
+module.exports = BadRequestError;
