+++ errors/ServerError.js
@@ -1,0 +1,8 @@
+class ServerError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 500;
+  }
+}
+
+module.exports = ServerError;
