+++ errors/TokenError.js
@@ -1,0 +1,8 @@
+class TokenError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 401;
+  }
+}
+
+module.exports = TokenError;
